requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
asyncpg==0.29.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic-settings==2.5.3
